backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
python-dotenv==1.0.1
fake-useragent==1.4.0
schedule==1.2.1
selenium==4.18.1
webdriver-manager==4.0.1

moviepy==1.0.3
numpy
